* add follow/unfollow * fix

+++ src/main/java/by/bsuir/service/dto/impl/EditCategoryServiceImpl.java
@@ -4,19 +4,22 @@ import by.bsuir.dto.category.SaveCategoryDto;
 import by.bsuir.dto.category.UpdateCategoryDto;
 import by.bsuir.entity.Category;
 import by.bsuir.entity.User;
-import by.bsuir.entity.enums.category.CATEGORY_TYPE;
 import by.bsuir.exception.BusinessException;
+import by.bsuir.exception.enums.ERROR_CODE;
 import by.bsuir.service.business.CheckCategoryService;
 import by.bsuir.service.business.SecurityService;
 import by.bsuir.service.dto.EditCategoryService;
 import by.bsuir.service.entity.CategoryService;
 import by.bsuir.service.entity.UserService;
 import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 
 import static by.bsuir.entity.enums.category.CATEGORY_TYPE.CUSTOM;
+import static by.bsuir.exception.enums.ERROR_CODE.CATEGORY_NOT_FOUND;
+import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 @Service
 public class EditCategoryServiceImpl implements EditCategoryService {
@@ -57,6 +60,10 @@ public class EditCategoryServiceImpl implements EditCategoryService {
         User user = userService.findByFirebaseId(getUid());
         Category category = categoryService.findById(categoryId);
 
+        if (!userService.existsByUserIdAndCategory(user.getId(), category)) {
+            throw new BusinessException(String.format("Category with id %s not found", categoryId), CATEGORY_NOT_FOUND, NOT_FOUND);
+        }
+
         user.getCategory().remove(category);
         categoryService.delete(category);
         return true;
@@ -66,7 +73,13 @@ public class EditCategoryServiceImpl implements EditCategoryService {
     @Transactional
     @Modifying
     public boolean update(UpdateCategoryDto updateCategoryDto) {
+        User user = userService.findByFirebaseId(getUid());
         Category category = categoryService.findById(updateCategoryDto.getId());
+
+        if (!userService.existsByUserIdAndCategory(user.getId(), category)) {
+            throw new BusinessException(String.format("Category with id %s not found", updateCategoryDto.getId()),
+                                                                                        CATEGORY_NOT_FOUND, NOT_FOUND);
+        }
 
         category.setName(updateCategoryDto.getName());
         category.setTextColor(updateCategoryDto.getTextColor());
